🚚(schemas) move core schemas to a dedicated file
Better separate core schema from utilities that can be used elsewhere to
prevent potential circular import issues.
